remove unused code and move engine to new file
