+++ bin/exec/src/main.rs
@@ -7,7 +7,6 @@ fn main() {
         .run(|builder, _| async move {
             let handle = builder
                 .with_types::<EthereumNode>()
-                // TODO: [now]: should we just use ethereum node and remove ivm node?
                 .with_components(EthereumNode::components().pool(IvmPoolBuilder::default()))
                 .with_add_ons(IvmAddOns::default())
                 .launch()
